Add dom; speed slider and reset button
Change title of html page to 'Tetris' from 'Code!'
Add reset button.
Add slider that impacts the speed of the game.
It works with inverse quadratic growth, because the default linear
growth has the first 20% being really important and the rest boring.
This way you can be less precise with your mouse to pick the option
or probably want.

Add reset button.