Fully work (I think) with population of 1 after this fix.
Genatic Algorithm:
	Create crossover - I think its important and copies the tutorial thing
	Create mutation - Copies tutorial thing and the mutate function that
			I already had. I probably should've just used that but I
			forogt that it existed.

After deciding where to go, the "player" tries to move their 1 move at a
time and turns 90 degrees at a time. The angle formula is hard and I
needed a piece of paper to devise it. I tried to explain it to my friend
and copy pasted my explanation.

Remove unnessessary code like
1. The one hot vector for moving the tetromino
that was needed for the NEAT Tetris but not the GA Tetris
(which is this one.)
2. Removed debug for drawing the board
3. Removed the ability to update the board without moving the tetrominoes down. This was implemented so the human could rotate the tetromino and see the changes in real time but that is no longer applicable.